tests: add tests for VP allowlist
